Add modelUtils to shareda and use it on snapshot subscriptions
