+++ client/src/routes/Session/hooks/useSubscribeToSession.ts
@@ -1,23 +1,28 @@
 import {useCallback} from 'react';
-import firestore, {
-  FirebaseFirestoreTypes,
-} from '@react-native-firebase/firestore';
-import {Session} from '../../../../../shared/src/types/Session';
+import firestore from '@react-native-firebase/firestore';
+import {Session, SessionData} from '../../../../../shared/src/types/Session';
+import {getData} from '../../../../../shared/src/modelUtils/firestore';
+import {getSession} from '../../../../../shared/src/modelUtils/session';
 
 const useSubscribeToSession = (sessionId: Session['id']) => {
   return useCallback(
-    (
-      onSnapshot: (
-        snapshot: FirebaseFirestoreTypes.DocumentSnapshot<FirebaseFirestoreTypes.DocumentData>,
-      ) => any,
-    ) => {
+    (onSnapshot: (session: Session) => any) => {
       const doc = firestore().collection('sessions').doc(sessionId);
 
-      const unsubscribe = doc.onSnapshot(onSnapshot, error =>
-        console.debug(
-          `Failed to subscribe to live session ${sessionId}`,
-          error,
-        ),
+      const unsubscribe = doc.onSnapshot(
+        snapshot => {
+          if (!snapshot.exists) {
+            return undefined;
+          }
+
+          const session = getSession(getData<SessionData>(snapshot));
+          onSnapshot(session);
+        },
+        error =>
+          console.debug(
+            `Failed to subscribe to live session ${sessionId}`,
+            error,
+          ),
       );
 
       return unsubscribe;
